refactor rqalpha_mod_sys_risk
